divides create-gauge
